rename: 채팅 소켓 함수명 변경
