+++ src/components/Drawer/Drawer.tsx
@@ -5,7 +5,7 @@ import styled from 'styled-components';
 import {IDrawer} from './Drawer.types';
 
 // TODO: 공통 컴포넌트 만들 예정
-const Drawer: React.FC<IDrawer> = ({isOpen, onClose, usersMap}) => {
+const Drawer: React.FC<IDrawer> = ({isOpen, onClose}) => {
   if (!isOpen) return null;
   const portalRoot = document.getElementById('drawer-root') as HTMLElement;
   return ReactDOM.createPortal(
