+++ src/contexts/ChatSocketContext.tsx
@@ -9,7 +9,7 @@ interface SocketState {
   prevMessages: PrevMessage;
   users: UserID;
 }
-const SocketContext = createContext<SocketState | null>(null);
+const ChatSocketContext = createContext<SocketState | null>(null);
 
 interface SocketProviderProps {
   id: string;
@@ -17,7 +17,7 @@ interface SocketProviderProps {
   children: React.ReactNode;
 }
 
-export const SocketProvider: React.FC<SocketProviderProps> = ({id, url, children}) => {
+export const ChatSocketProvider: React.FC<SocketProviderProps> = ({id, url, children}) => {
   const [socket, setSocket] = useState<Socket | null>(null);
   const [messages, setMessages] = useState<Message[]>([]);
   const [prevMessages, setPrevMessages] = useState<PrevMessage>({messages: []});
@@ -27,6 +27,13 @@ export const SocketProvider: React.FC<SocketProviderProps> = ({id, url, children
     const newSocket = io(url, {
       extraHeaders: authHeaders(),
     });
+    newSocket.off('connect');
+    newSocket.off('disconnect');
+    newSocket.off('message-to-client');
+    newSocket.off('messages-to-client');
+    newSocket.off('users-to-client');
+    newSocket.off('join');
+    newSocket.off('leave');
 
     newSocket.on('connect', () => {
       console.log('Socket connected:', newSocket.id);
@@ -39,12 +46,6 @@ export const SocketProvider: React.FC<SocketProviderProps> = ({id, url, children
     newSocket.on('disconnect', reason => {
       console.log('Socket disconnect:', reason);
     });
-
-    newSocket.off('message-to-client');
-    newSocket.off('messages-to-client');
-    newSocket.off('users-to-client');
-    newSocket.off('join');
-    newSocket.off('leave');
     // 메시지 수신
     newSocket.on('message-to-client', (data: Message) => {
       setMessages(prevMessages => [...prevMessages, data]);
@@ -82,7 +83,6 @@ export const SocketProvider: React.FC<SocketProviderProps> = ({id, url, children
 
     return () => {
       if (newSocket.connected) {
-        console.log('소켓 연결종료');
         newSocket.close();
       }
     };
@@ -95,11 +95,11 @@ export const SocketProvider: React.FC<SocketProviderProps> = ({id, url, children
     users,
   };
 
-  return <SocketContext.Provider value={contextValue}>{children}</SocketContext.Provider>;
+  return <ChatSocketContext.Provider value={contextValue}>{children}</ChatSocketContext.Provider>;
 };
 
 export const useSocketContext = () => {
-  const context = useContext(SocketContext);
+  const context = useContext(ChatSocketContext);
   if (!context) {
     throw new Error('useSocketContext must be used within a SocketProvider');
   }
